% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.coxph}
\alias{as.coxph}
\alias{as.coxph.rcv.glmnet}
\alias{as.coxph.coxnet}
\alias{as.cph}
\alias{as.cph.rcv.glmnet}
\alias{as.cph.coxnet}
\title{Convert coxnet objects}
\usage{
as.coxph(object, ...)

\method{as.coxph}{rcv.glmnet}(object, s, x, y, ...)

\method{as.coxph}{coxnet}(object, s, x, y, ...)

as.cph(object, ...)

\method{as.cph}{rcv.glmnet}(object, s, x, y, ...)

\method{as.cph}{coxnet}(object, s, x, y, ...)
}
\arguments{
\item{object}{object that should be converted.}

\item{\dots}{further arguments passed to the underlying function
(\code{survival::coxph} or \code{rms::cph}).}

\item{s}{\code{character}/\code{numeric}, value(s) of the penality parameter \code{lambda}.}

\item{x}{\code{matrix}, as in \code{glmnet}.}

\item{y}{response as in \code{glmnet}.}
}
\description{
These functions convert \code{coxnet} objects into popular survival objects like
\code{survival::coxph} or \code{rms::cph}.
}
\seealso{
\code{\link[glmnet:glmnet]{glmnet::glmnet()}}

\code{\link[rms:cph]{rms::cph()}}
}
\author{
Sebastian Gibb
}
